Refactor(#175): 모임 생성/수정 시 쿠키가 자동으로 전송되도록 수정
